% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a1Coefficients}
\alias{a1Coefficients}
\title{Caclulate the coefficients a_1 in the expression for K_1}
\usage{
a1Coefficients(n)
}
\arguments{
\item{n}{an integer}
}
\description{
Caclulate the coefficients a_1 in the expression for K_1
}
\keyword{internal}
