% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_K.R
\name{a3Coefficients}
\alias{a3Coefficients}
\title{Caclulate the coefficients a_3 in the expression for K_3}
\usage{
a3Coefficients(n)
}
\arguments{
\item{n}{an integer}
}
\description{
Caclulate the coefficients a_3 in the expression for K_3
}
\keyword{internal}
