% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateK}
\alias{calculateK}
\title{Caclulate K_p for a BSS process for a given value of p}
\usage{
calculateK(p, alpha)
}
\arguments{
\item{p}{an integer - the power to use for K_p}

\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate K_p for a BSS process for a given value of p
}
\keyword{internal}
