% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_K.R
\name{calculateK1}
\alias{calculateK1}
\title{Caclulate K_1 for a BSS process}
\usage{
calculateK1(alpha)
}
\arguments{
\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate K_1 for a BSS process
}
\keyword{internal}
