% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateK2}
\alias{calculateK2}
\title{Caclulate K_2 for a BSS process}
\usage{
calculateK2(alpha)
}
\arguments{
\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate K_2 for a BSS process
}
\keyword{internal}
