% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateK3}
\alias{calculateK3}
\title{Caclulate K_3 for a BSS process}
\usage{
calculateK3(alpha)
}
\arguments{
\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate K_3 for a BSS process
}
\keyword{internal}
