% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_K.R
\name{calculateK4}
\alias{calculateK4}
\title{Caclulate K_4 for a BSS process}
\usage{
calculateK4(alpha)
}
\arguments{
\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate K_4 for a BSS process
}
\keyword{internal}
