% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateK}
\alias{estimateK}
\title{Estimate K_p for a BSS process, for a given power p}
\usage{
estimateK(Y, p)
}
\arguments{
\item{Y}{a vector of observations of the BSS process}

\item{p}{an integer - the power to use for K_p}
}
\description{
Estimate K_p for a BSS process, for a given power p
}
\keyword{internal}
