% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{gammaKernelCorrelation}
\alias{gammaKernelCorrelation}
\title{Autocorrelation function for the gamma kernel}
\usage{
gammaKernelCorrelation(alpha, lambda, h)
}
\arguments{
\item{alpha}{the smoothness parameter, alpha, for the gamma kernel.}

\item{lambda}{the exponent parameter, lambda, for the gamma kernel.}

\item{h}{the lag to calculate the autocorrelation at.}
}
\value{
The function returns the autocorrelation for the gamma kernel with
parameters \code{alpha} and \code{lambda} at lag \code{h}.
}
\description{
\code{gammaKernelCorrelation} calculates the value of the gamma kernel autocorrelation function
directly using the analytic expression.
}
