% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{gammaKernelTau}
\alias{gammaKernelTau}
\title{Scale factor for the gamma kernel}
\usage{
gammaKernelTau(n, alpha, lambda)
}
\arguments{
\item{n}{a positive integer indicating the number of observations per unit of time.}

\item{alpha}{the smoothness parameter, alpha, for the gamma kernel.}

\item{lambda}{the exponent parameter, lambda, for the gamma kernel.}
}
\value{
The function returns the scale factor (tau_n) for the gamma kernel with
parameters \code{alpha} and \code{lambda}, observed at frequency \code{n} per unit of time.
}
\description{
\code{gammaKernelTau} evaluates the scale factor tau_n for the gamma kernel using the
exact expression derived from the covariance function.
}
