% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{gammaKernelTauAsymptotic}
\alias{gammaKernelTauAsymptotic}
\title{Asymptotic scale factor for the gamma kernel}
\usage{
gammaKernelTauAsymptotic(n, alpha)
}
\arguments{
\item{n}{a positive integer indicating the number of observations per unit of time.}

\item{alpha}{the smoothness parameter, alpha, for the gamma kernel.}
}
\value{
The function returns an approximation for the scale factor (tau_n) for the gamma kernel with
smoothness parameter \code{alpha}, observed at frequency \code{n} per unit of time, using the asymptotic
expression for the scale factor.
}
\description{
Asymptotic scale factor for the gamma kernel
}
