% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{powerKernelCorrelation}
\alias{powerKernelCorrelation}
\title{Autocorrelation function for the power law kernel}
\usage{
powerKernelCorrelation(alpha, beta, h)
}
\arguments{
\item{alpha}{the smoothness parameter, alpha, for the power law kernel.}

\item{beta}{the exponent parameter, beta, for the power law kernel.}

\item{h}{the lag to calculate the autocorrelation at.}
}
\value{
The function returns the autocorrelation for the power law kernel with
parameters \code{alpha} and \code{beta} at lag \code{h}.
}
\description{
\code{powerKernelCorrelation} calculates the value of the power law kernel autocorrelation function
directly using numerical integration for the numerator (the covariance term) and the analytic expression
for the denominator (variance term).
}
