% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{powerKernelTau}
\alias{powerKernelTau}
\title{Scale factor for the power law kernel}
\usage{
powerKernelTau(n, alpha, beta)
}
\arguments{
\item{n}{a positive integer indicating the number of observations per unit of time.}

\item{alpha}{the smoothness parameter, alpha, for the power law kernel.}

\item{beta}{the exponent parameter, beta, for the power law kernel.}
}
\value{
The function returns the scale factor (tau_n) for the power law kernel with
parameters \code{alpha} and \code{beta}, observed at frequency \code{n} per unit of time.
}
\description{
\code{powerKernelTau} evaluates the scale factor tau_n for the power law kernel using
numerical integration for the covariance term, and exact evaluation for the variance term.
}
