% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_functions.R
\name{realisedPowerVariation}
\alias{realisedPowerVariation}
\title{Realised power variation}
\usage{
realisedPowerVariation(Y, p)
}
\arguments{
\item{Y}{a vector of observations of a BSS process.}

\item{p}{the power variation to be calculated.}
}
\value{
The function returns the sum of the pth powers of the absolute first differences of the BSS process.
}
\description{
\code{realisedPowerVariation} calculates the realised power variation for a BSS process, which is then
used as a component of estimating the accumulated volatility process.
}
