% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_K.R
\name{rhoFractionGaussian}
\alias{rhoFractionGaussian}
\title{Caclulate the correlation of a fractional Gaussian - needed in the calculation of K}
\usage{
rhoFractionGaussian(j, alpha)
}
\arguments{
\item{j}{an integer}

\item{alpha}{a float, the smoothness parameter of the BSS process}
}
\description{
Caclulate the correlation of a fractional Gaussian - needed in the calculation of K
}
\keyword{internal}
